feat(post): add post endpoints (#25)

+++ src/api-gateway/schema/common.go
@@ -1,7 +1,7 @@
 package schema
 
 type PaginationResponse struct {
-	Offset  int32 `json:"offset"`
+	Offset  int64 `json:"offset"`
 	Limit   int32 `json:"limit"`
 	Records int32 `json:"records"`
 }
